Refactor OTVision config parsing to support different types of input sources
This commit will:
- Move config parsing logic to dedicated ConfigParser class
- Make ConfigParser abstract class acting as a template to extending
subclasses supporting different types of input sources
- Move entire Config class to its own module in the application layer
- Make DetectBuilder an abstract class to enable implementation
of different types of DetectBuilders with different input sources
