+++ OTVision/application/detect/update_detect_config_with_cli_args.py
@@ -1,5 +1,5 @@
+from OTVision.application.config import Config, DetectConfig, YoloConfig, _LogConfig
 from OTVision.application.detect.get_detect_cli_args import GetDetectCliArgs
-from OTVision.config import Config, DetectConfig, YoloConfig, _LogConfig
 from OTVision.domain.cli import DetectCliArgs
