+++ OTVision/domain/object_detection.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Generator
 
-from OTVision.config import DetectConfig
+from OTVision.application.config import DetectConfig
 from OTVision.domain.detection import DetectedFrame
 from OTVision.domain.frame import Frame
