+++ OTVision/plugin/config_parser/streaming_source.py
@@ -1,0 +1,6 @@
+from OTVision.application.config_parser import ConfigParser
+
+
+class StreamingSourceConfigParser(ConfigParser):
+    def parse_sources(self, sources: list[str]) -> list[str]:
+        return sources
